EDU-1061 Merge profile and account tabs under settings tab
