+++ src/components/account-settings-tab.yml
@@ -1,0 +1,50 @@
+info:
+  en: Here you can update your account information. The personal information is publicly accessible through your <0>profile page</0>.
+  de: Hier kannst du deine Konto-Informationen bearbeiten. Die Persönlichen Daten sind auf deiner <0>Profilseite</0> öffentlich sichtbar.
+personalInfoHeadline:
+  en: Personal info
+  de: Persönlichen Daten
+changePicture:
+  en: Change profile picture
+  de: Profilbild ändern
+avatarTitle:
+  en: Gravatar profile picture
+  de: Gravatar Profilbild
+organization:
+  en: Organization
+  de: Organisation
+introduction:
+  en: Introduction
+  de: Vorstellung
+avatarDescription:
+  en: |
+    Profile pictures are managed through the <0>Gravatar</0> internet service.
+    A Gravatar is a globally recognized avatar, which is linked to your email address.
+    This service can then be used by other websites to display a profile picture for a given email address.
+    Please note that our website has no influence on the use of your data by this service.
+    If you agree, you can create a profile <0>here</0>.
+  de: |
+    Profilbilder werden durch den Internetdienst <0>Gravatar</0> verwaltet.
+    Ein Gravatar ist ein global verfügbarer Avatar (Globally Recognized Avatar), welcher mit deiner E-Mail-Adresse verknüpft ist.
+    Dieser Dienst kann dann von anderen Webseiten benutzt werden, um ein Profilbild zu einer E-Mail-Adresse anzuzeigen.
+    Beachte bitte, dass unsere Website keinen Einfluss auf die Verwendung deiner Daten bei diesem Dienst hat.
+    Wenn du damit einverstanden bist, kannst du <0>hier</0> ein Profil erstellen.
+credentialsHeadline:
+  en: Credentials
+  de: Zugangsdaten
+resetPassword:
+  en: Reset password
+  de: Kennwort wiederherstellen
+updateSuccessMessage:
+  en: Your account was updated successfully.
+  de: Dein Konto wurde erfolgreich aktualisiert.
+passwordResetEmailSent:
+  en: The password reset link has been sent to {email}
+  de: Der Link zum Zurücksetzen des Kennworts wurde an {email} gesendet
+closeAccountDetails:
+  en: All your public contributions (documents and files uploaded to the public storage) will remain on the website and your name will still be displayed as author or contributor. All other data (personal information, rooms you own and files uploaded to your private storage) will be removed. You will no longer be able to log into the website.
+  de: Alle deine öffentlichen Beiträge (Dokumente und Dateien, die du in den öffentlichen Speicher hochgeladen hast), bleiben auf der Website. Dein Anzeigename wird weiterhin als Autor oder Bearbeiter angezeigt. Alle anderen Daten (persönliche Informationen, Räume, die du besitzt, und Dateien, die du in den privaten Speicher hochgeladen hast), werden gelöscht. Du kannst dich dann nicht mehr auf der Website anmelden.
+closeAccount:
+  en: Close account
+  de: Konto schließen
+
