--- src/components/account-tab.less
@@ -1,21 +1,0 @@
-.AccountTab {}
-
-.AccountTab-tabInfo {
-  .m-info-card;
-}
-
-.AccountTab-headline {
-  .m-section-headline;
-}
-
-.AccountTab-section {
-  .m-section-content-narrow;
-
-  &.AccountTab-section--wide {
-    max-width: none;
-  }
-}
-
-.AccountTab-closeAccountButton {
-  margin-top: 15px;
-}
