+++ src/components/pages/dashboard.yml
@@ -4,12 +4,9 @@ activitiesTabTitle:
 favoritesTabTitle:
   en: Favorites
   de: Favoriten
-profileTabTitle:
-  en: Profile
-  de: Profil
-accountTabTitle:
-  en: Account
-  de: Konto
+settingsTabTitle:
+  en: Settings
+  de: Einstellungen
 roomsTabTitle:
   en: Rooms
   de: Räume
