--- src/components/profile-tab.less
@@ -1,33 +1,0 @@
-.ProfileTab {}
-
-.ProfileTab-tabInfo {
-  .m-info-card;
-}
-
-.ProfileTab-headline {
-  .m-section-headline;
-}
-
-.ProfileTab-section {
-  .m-section-content;
-
-  @media @edu-media-phone {
-    padding-left: 0;
-    padding-right: 0;
-  }
-}
-
-.ProfileTab-input {
-  max-width: 50%;
-
-  @media @edu-media-phone {
-    max-width: none;
-  }
-}
-
-.ProfileTab-avatar {
-  display: grid;
-  grid-auto-flow: row;
-  gap: @edu-control-spacing-base;
-  margin-bottom: 20px;
-}
