--- src/components/profile-tab.yml
@@ -1,35 +1,0 @@
-info:
-  en: Here you can provide additional information about yourself. This information is publicly accessible through your <0>profile page</0>.
-  de: Hier kannst du zusätzliche Informationen zu deiner Person angeben. Die Daten sind auf deiner <0>Profilseite</0> öffentlich sichtbar.
-personalInfoHeadline:
-  en: Personal info
-  de: Persönlichen Daten
-changePicture:
-  en: Change profile picture
-  de: Profilbild ändern
-howToChangePicture:
-  en: How can I change my profile picture?
-  de: Wie ändere ich mein Profilbild?
-organization:
-  en: Organization
-  de: Organisation
-introduction:
-  en: Introduction
-  de: Vorstellung
-updateSuccessMessage:
-  en: Your profile was updated successfully.
-  de: Dein Profil wurde erfolgreich aktualisiert.
-avatarDescription:
-  en: |
-    Profile pictures are managed through the <0>Gravatar</0> internet service.
-    A Gravatar is a globally recognized avatar, which is linked to your email address.
-    This service can then be used by other websites to display a profile picture for a given email address.
-    Please note that our website has no influence on the use of your data by this service.
-    If you agree, you can create a profile <0>here</0>.
-  de: |
-    Profilbilder werden durch den Internetdienst <0>Gravatar</0> verwaltet.
-    Ein Gravatar ist ein global verfügbarer Avatar (Globally Recognized Avatar), welcher mit deiner E-Mail-Adresse verknüpft ist.
-    Dieser Dienst kann dann von anderen Webseiten benutzt werden, um ein Profilbild zu einer E-Mail-Adresse anzuzeigen.
-    Beachte bitte, dass unsere Website keinen Einfluss auf die Verwendung deiner Daten bei diesem Dienst hat.
-    Wenn du damit einverstanden bist, kannst du <0>hier</0> ein Profil erstellen.
-
