+++ src/styles/main.less
@@ -31,8 +31,6 @@
 @import '../components/byte-input.less';
 @import '../components/item-panel.less';
 @import '../components/mini-pager.less';
-@import '../components/account-tab.less';
-@import '../components/profile-tab.less';
 @import '../components/collapsible.less';
 @import '../components/pages/index.less';
 @import '../components/pages/login.less';
@@ -77,6 +75,7 @@
 @import '../components/default-page-footer.less';
 @import '../components/default-page-header.less';
 @import '../components/blocked-login-error.less';
+@import '../components/account-settings-tab.less';
 @import '../components/pages/reset-password.less';
 @import '../components/media-player/timeline.less';
 @import '../components/default-page-template.less';
